Plot now plots one of the multiple paths
There are no runtime errors, but only one of the paths is being plotted.

This closes #11. Because all that remains is GUI related to plotting.

But the priority of #27 is now high.

in MPLViewer.plot:
* label must be pased as **kwargs to axes.plot

in sdp_project_demo.plot:
* for path in paths
* paths list is in form [path, "plannerName"]

TODO: add a legend to the MPLViewer using the labels.
TODO: double check all Qt Signal and Socket connections
TODO: Investigate why multiple paths aren't being plotted.