Removed Connections between MPLViewer and Bounds
This Closes #27

The plot viewer should not be able to set the bounds. Only the user 
should set the bounds. This fixes the connections issues from #27. 

Added getBounds function to boundsWidget. 

The bounds need to be passed to MPLViewer every time it plots.
Otherwise MatPlotLib will try to use smart automatic axis values.

added setBounds in sdp_project_demo. bounds are set when solve
calls setupPlanners. Solve enables the plot button. The bounds are
used when the plot button is clicked.